Final version of the RTEMS compiler
Together with Embedded Brains, multiple issues relating to
failing RTEMS FPU tests have been addressed over the last
two weeks - by executions of the RTEMS suite on both the
GR712 and the GR740. This latest version of the toolchain
which can be rebuilt, by those who care, via our build
script at https://goo.gl/Nrb9Y3 - and includes support for:

    - compiling Ada code
    - native FPU targetting
    - SMP (i.e. takes advantage of all cores of GR712/GR740)

The only side-effect is that the Leon3 QEMU simulator now
fails to load the binaries - but we can't let that stop
moving to a new release.

Executive summary:
    RSB commit:   e2952bb185c1f40027caa76cfd9e4a45b17a8886
    RTEMS commit: 96ce1ec743a1fcf27593ee72cf1695d9eb0290de