Add ability to install add-ons
There are many things that are not supposed to be part
of the pre-made VM - including custom BSPs for different
targets, build toolchains, etc.

This folder will allow said functionality - for now, it
allows installation of non-SMP BSPs for Leon2 (non-FPU)
and Leon3 (FPU), that work fine with the bundled QEMU
emulator.